requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
openai==1.40.0
httpx==0.27.2
Flask-Limiter==3.8.0
